Updated releng/autorelease Project: ovsdb  444b68f6f7637d9b4eb7e4f07c9ff6ced4e8ce76
Bug 3831 - add stt interface type to model

On a build from openvswitch master branch, there was an
interface type 'stt' which was not included in the ovsdb.yang
model for the southbound MD-SAL.  When an OVSDB connection
was made to this OVS node, the operational md-sal failed
to populate.

Adding the 'stt' interface type to the model is a quick
fix.  A better fix would ignore types not present in the model.

Change-Id: Idf259998f73d43be587d19de73794a174d4d81d3
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>