Updated releng/autorelease Project: ovsdb  c525fd2b379171af656ffd060d9e520ab4e3854e
Bug 3831 - ignore unknown OVS interface and datapath types

Ighore OVSDB interface or datapath types which are not
present in the MD-SAL southbound model (but are present
in the OVS nodes open_vswitch table iface-types or
datapath-types columns).

Tested by making a build of ODL which removed supported
types from the ODL code to varify the log warnings
were created.

Change-Id: I05e689fa2b2dc906e3083e4fc44ba60abdb1d352
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>