Updated releng/autorelease Project: integration  b9deb5f32d7fecb628ea7aaa0f568bfb5a216941
Update clustering tests for Continuous Integration

Moved clustering tests to controller directory.

Replaced ${PORT} with ${RESTCONFPORT} in call to Wait For Controller Up.

This commit is a re-try of the following commit:
https://git.opendaylight.org/gerrit/#/c/23001/

It adds KEYFILE_PASS to variables.py
and then puts KEYFILE_PASS into
Run Command On Remote System.

It also contains the following from the
previous commit attempt:
    Change global variables to match those used in CI
    All ssh sign-ins now use certificates rather than passwords
    Remove Sleeps where possible.
    Move repeated variables into variables.py.
    Move repeated keywords into libraries.

Change-Id: Iacb328c89abe35d8c57eac4b4d081ff219b2e8b5
Signed-off-by: Phillip Shea <phillip.shea@hp.com>