Updated releng/autorelease Project: persistence  a9b089561bfc5971cba2261a9a850dbf3f0c3c75
Add persistence-common as dependency of features-persistence

Distribution jobs are currently failing SingleFeatureTest,
the test is unable to resolve persistence-common artifact.
This change adds that artifact as an explicit dependency
of features-persistence artifact.

Only odl-persistence-api is currently present in integration,
so this change does not add other dependencies of features-persistence
(such as persistence-jpa or test-app-*).

Change-Id: Ia5c336c3be4b6ebf7b1f15f6081ce95c7b50cf51
Signed-off-by: Vratko Polak <vrpolak@cisco.com>