Updated releng/autorelease Project: aaa  de7091c0dd852874111140e2dede264ff3a80b69
Bug 2923 Restarting the controller causes multiple insertions users

Changes criteria of StoreBuilder.init() so it is called when
"idmlight.db.mv.db" does not exist.  Previously, the code checked for the
existence of "idmlight.db".  This convention changed with the addition of
H2 as the data store backing AAA data.

Change-Id: I54dcc896aec57b8b89b3be7dfe0c50a7973d49ce
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>