Updated releng/autorelease Project: aaa  79a0b26a955a8ddde613393bb61c6ca0c72e4a41
Bug 2923 Restarting the controller causes multiple insertions users

Changes criteria of StoreBuilder.init() so it is called when
"idmlight.db.mv.db" does not exist.  Previously, the code checked for the
existence of "idmlight.db".  This convention changed with the addition of
H2 as the data store backing AAA data.

Change-Id: Ifa15ce89b414bbbb32b34d2ca38a3504921d8e27
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>