Updated releng/autorelease Project: odlparent  b4d05f744578e543d3b6dc5685468342436b6cdf
BUG-3964 Enabled the generation of sources and javadocs artifacts for all projects. The relevant
plugins were already present in the <pluginManagement> section. These were added to the execution
list as well.

Tested this change by building the controller, openflowplugin and l2switch projects. The sources and
javadocs jar files were generated for the components.

Change-Id: I493c7479ecac189b4eec5e4190c707264219cf1b
Signed-off-by: Shaleen Saxena <ssaxena@brocade.com>