Updated releng/autorelease Project: neutron  d86c3a22a2431cb1b4c974ce36fed239938e83ec
Fix code duplication in report to sonar

The patch that got IT reporting working had the nasty
side effect of doubling the LoC reported by sonar
for the module. This patch rearranges things to
fix that and report IT results for each of the modules,
rather than from the IT module.

Change-Id: Ied75efa67ce8843c6650d3a04a333586bf35c747
Signed-off-by: Ryan Moats <rmoats@us.ibm.com>