Updated releng/autorelease Project: lispflowmapping  6e425e7db05425792380c0ae5e0bea028b25a111
Move HashMapDb to its own bundle

HashMapDb is a simple, in memory implementation of ILispDAO, the
interface that defines the API for the interaction with the Map-Server's
mappings database. Until now, HashMapDb was part of
mappingservice.implementation but with this commit it is moved to its
own mappingservice.inmemorydb bundle.

An important feature is that a ILispDAO implementation is now injected
into mappingservice.implementation at runtime through config-subsystem.
As a result, InMemoryDb can be easily swapped with other implementations.

Change-Id: I29af97b831414eecbbf3c29e6fbec086a627a770
Signed-off-by: Florin Coras <fcoras@cisco.com>