Updated releng/autorelease Project: bgpcep  d99e924ff10bb7652ef12e7989540b9ffecb2dd3
BUG-3888 : fix comparing ASnumbers

In order to allow BGP connection, expected remote AS (configured in ODL)
needs to match AS number received in peer Open message. Added expected
remote AS to BGP Session Preferences to allow for this comparison.

Moved validate method from client validator to registry that is used
anytime a peer connects.

This commit is intended to be cherry-picked. Refactoring commit
will be followed up in master branch.

Change-Id: Ie8bfd8b51ee7654b284fe0b0d4ce1db5e365d55b
Signed-off-by: Dana Kutenicsova <dkutenic@cisco.com>