Updated releng/autorelease Project: integration  10105981349e15d5722685a66a931f7fc71a10a3
Currently, the addCar and addPerson methods create multiple cars/people in tight loops. These
loops do not check the results of the individual post operations. In case of cluster failures,
each post operation may take up to 5 minutes to timeout. Hence, these loops could take many
hours to complete.

The fix is to add an additional 'expected' parameter to these methods. This is a list of possible
results expected by the post operation. If any post operation does not match any of the 'expected'
values, then the loop will exit and raise an exception. This way the loop will wait for only
few minutes instead of hours.

Please note that this 'expected' parameter is optional. If it is not provided, then the old
behavior is preserved.

The other changes are to pass the expected values from various AddCar and AddPerson test cases.

Change-Id: I7375979de90f64c3ce70cfe4f3f6634d0aeac8e2
Signed-off-by: Shaleen Saxena <ssaxena@brocade.com>