Updated releng/autorelease Project: neutron  5bcf40a1087adf8b232c07407e2c209a4f65bfc2
Make MDSAL delete synchronous

Before this patch the future returned by read and put
transactions was wrapped in a get to render the result
synchronous (this is necessary to support Neutron).
Apply this pattern to the future returned by the
delete transaction.

This allows the negative get network test post delete
to work again, so re-enable it.

Change-Id: I2443a07869114ae466549de5f5f03132f5d3bdf1
Signed-off-by: Ryan Moats <rmoats@us.ibm.com>