Updated releng/autorelease Project: ovsdb  8e668f87a71ad583d620e2c5b8e51996e9a74bde
Indicate that RefType values need to be lower-case

Sonar complains that RefType's values should be upper-cased (as
constants), but since they match JSON values they need to be lower-case.

While we're at it, drop an unnecessary null declaration, remove the
redundant static modifier from RefType, and merge two if statements.

Change-Id: I98c7baf63cf42784ba40416bff96259ae8327084
Signed-off-by: Stephen Kitt <skitt@redhat.com>