Updated releng/autorelease Project: bgpcep  c27ebb404476d713300ade468818a76e16401402
Fixed possible race condition when using BGP as speaker.

There is a short period of time when channel is active, but ODL
is yet to send Open and the peer sends Open. In this case we
would throw a Notify exception (unexpected messages) which is not
correct. Expect also Open when in IDLE state.

Change-Id: I7e4406ab1d8c29fac8ca56d9a6c797bb00349504
Signed-off-by: Dana Kutenicsova <dkutenic@cisco.com>