Updated releng/autorelease Project: lispflowmapping  ecc3f75c6ff6485eaedfe3938351b24effb4bbec
Bug 4077: Send SMR when deleting a Src/Dst mapping

When we had a general mapping and within that a src/dst mapping,
deleting the src/dst mapping was not generating an SMR to the
subscribers of said  mapping. This commit fixes that.

Change-Id: I199af9f8558a8ebbf4a100f224c570893f548c20
Reported-by: Vina Ermagan <vermagan@cisco.com>
Signed-off-by: Lorand Jakab <lojakab@cisco.com>
(cherry picked from commit 551b102f4ae329dca674509120acbecfd035b716)