Updated releng/autorelease Project: ovsdb  38eef6358b581a680ac14f3d5fb6aa6203e95bc1
Fix for defect 4135

https://bugs.opendaylight.org/show_bug.cgi?id=4135

If integration bridge or external bridge exist prior to
connecting to ODL, then the pre-existing bridge is not added.
But the side effect is that the controller is not set on
this bridge.
Current code checks to see if the bridge already exists on the node. This
patch adds a check to see if the bridge exists on CONFIG datastore. If the
bridge does not exist on config, it indicates that the bridge existed prior
to ODL connection, and it adds the bridge.

(cherry picked from commit 920e5d6676bbcd7204dd077f5e19a139edb8b57a)

Change-Id: Iddf0e8463585346c06f86f0924ee0e3e5b8bdb31
Signed-off-by: Sharad Mishra <sharad.d.mishra@intel.com>