Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

Thanks Stephen, I'm not an expert in xmlstartlet but from what I've seen from this updated script it seems the changes should be ok. We don't need it to cleanup the pom.xml since none of the output is to be committed back to the projects but there's no harm in it if it results in the same xml.

Let's get this merged and see if it breaks anything in the daily builds.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=+1
Label: Code-Review=+2