Updated releng/autorelease Project: ovsdb  fee81ff402d6e6944ebf3f76be599c2e48ef9c5b
Fix for BUG 4067 (LBaaS flows don't get installed in the switch.)

LbaaS is installing the OF Flows in opendaylight-inventory CONFIG data store with NodeId stored in OVSDB topology data tree, but openflow plugin doesn't have any node with this node Id as they are based on DPID. 

Other NetVirt Provider Services map the OVSDB node Id to Southbound DPID, but LBaaS is not doing the mapping.

Fix is to get the corresponding DPID and create the Node Id that can be understood by the Openflow plugin and install the flow.


Change-Id: I485963b06da82c657f098d59e3e230fb6d381983
Signed-off-by: Arun S. Yerra <arunsarat@yahoo.com>
Fixing Test Failures with the Change# 24618

Change-Id: Id59b702329e9b1c8c14629e154d1c3e7a7f4fd88
Signed-off-by: Arun S. Yerra <arunsarat@yahoo.com>