Updated releng/autorelease Project: integration  f8423a1fb1e7d9f6f2ba2b2fab0810fea5667c87
${starttime} is a local variable, so we need to pass it to "Save Setup Time"

this is not seen in older versions of robotframe work because variable
scoping was not as strict.  In the new RF2.9, local variables do not
"leak".  This is probably a problem in many places, but I found this one
for now.  It breaks on my local RF2.9 environment.  Works with this
patch and should not have any problems upstream with older RF versions

Change-Id: I58e3f5b9ce1f71a86cf6bac8ca4c5e359e82832c
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>