Updated releng/autorelease Project: integration  08640680725f963de11918ea587f77186216cff9
Switched to "Get|Put|Delete" from "[Get|Put|Delete]_Request"

The "Get_Request" command has a different format and the
format differs with different versions of Robot Framework.
The "Get" is deprecated but is the same across different
Robot Framework versions so to fix the problem the test was
switched to use "Get". The same is true for "Put_Request"
versus "Put" and "Delete_Request" versus "Delete".

Change-Id: I0e28024e3384212db79520296ace3dd7a3dc95e9
Signed-off-by: Jozef Behran <jbehran@cisco.com>