Updated releng/autorelease Project: integration  1de8438cbdb428481bc06e167a78d350fbe565c7
Fixed failure due to a too sensitive command

There is a command in the "Verify Keyword Does Not Fail
Within Timeout" keyword which tried to catch the "timeout
failure" error but the problem is that the "timeout failure"
error looks different on different versions of Robot
Framework. The result is that the test fails on any Robot
Framework version except the one that it was debugged on.
The fix simply reduces the sensitivity of the command to
"any error is ok with me".

Change-Id: I4ecd7b5f7c2ea7bc070313dc69b3a7c7d3f96908
Signed-off-by: Jozef Behran <jbehran@cisco.com>
(cherry picked from commit cef1ea4aea0b0d2efec7e897011d15e3117cc02b)