Updated releng/autorelease Project: integration  170605616cf9d3786ce174d288898654902c8fdf
Updated the MessageError to produce less confusing messages

The first change is that there is a colon between the text of
the error message and the message hexdump and the second
change is that when the message is empty, then string "(empty
message)" is placed there instead of an empty string. These
changes shall reduce user's confusion.

Change-Id: I46deb344797fc6949d28fdccfb24f60c45e9115a
Signed-off-by: Jozef Behran <jbehran@cisco.com>
(cherry picked from commit 1355a04ffb866ce8eb252cd029e6107ed42b08ad)