Updated releng/autorelease Project: aaa  0016eecb852d455ff59b217bfb59a9dbe3bf5727
Bug 4146 odl-aaa-authz does not work for Write or ReadWrite Transactions

This patch modifies the authorization model by defining ActionTypes that are
applicable to the existing DOMDataReadOnlyTransaction,
DOMDataWriteOnlyTransaction and DOMDataReadWriteTransaction.  The ActionType
enum is utilized to add authorization to the AuthzWriteOnlyTransaction and
AuthzReadWriteTransaction.  Utility methods for ensuring authorization are
added to the AuthzServiceImpl.

Change-Id: Ia5c67946160e3dc0a074e616fbbb980e5eeafec1
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>