Updated releng/autorelease Project: integration  c46fcc035e5f69f2f5812c404f544952c2147235
Keyword to log messages to controller's karaf log

The keyword "Log Message To Controller Karaf" sends the
specified text to the Karaf log. It makes initialization
on its first call so no special initialization is necessary
and it does not drain any resources from suites that don't
use it. This keyword is not "cluster friendly" as it always
sends the message to the controller which is specified using
the "CONTROLLER" Robot variable.

Change-Id: I07417c242b5ad0e8fea2ee9910d2b44764a02ade
Signed-off-by: Jozef Behran <jbehran@cisco.com>
(cherry picked from commit ea4f231d21eb99bb545ebaef54556198103f58d5)