Updated releng/autorelease Project: yangtools  a25259656891f17f0851fe20b0d5af290ea78b29
Bug4231 - Yang tools failing to parse Augmentations under "uses" clause.

The problem occurs when uses-augment is nested in another Augment statement.
In this case the parent node of uses node is not subclass of SchemaNode
but it is AugmentationSchemaNode (it is not subclass of SchemaNode) and in
consequence class cast exception occurs.

(cherry picked from commit d417276b1c26ab5029b381657fd0be810970e572)
Change-Id: I8f58e3d326bb8a8dab9a187a333cd7b027223d75
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>