Updated releng/autorelease Project: yangtools  d417276b1c26ab5029b381657fd0be810970e572
Bug4231 - Yang tools failing to parse Augmentations under "uses" clause.

The problem occurs when uses-augment is nested in another Augment statement.
In this case the parent node of uses node is not subclass of SchemaNode
but it is AugmentationSchemaNode (it is not subclass of SchemaNode) and in
consequence class cast exception occurs.

Change-Id: I674392889e561f357297fa220051e8c3728f537d
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>