Updated releng/autorelease Project: ovsdb  4e22dfd3f72cb6b02022ea22c62147f7240074db
Switch from Felix to config-driven dependencies

The library module was regenerated using
opendaylight-startup-archetype 1.1.0-SNAPSHOT.

The ovsdb version is now 1.2.1-SNAPSHOT to make sure none of the
old 1.2.0-SNAPSHOT artefacts are used.

The library-dependent integration tests (at an OSGi level) are now in
library-it, apart from the plugin tests which are just ignored since
they're going away.

Change-Id: Ic6fb1c14f98b1557b9c87a702b20ef8a3173daf7
Signed-off-by: Stephen Kitt <skitt@redhat.com>