Updated releng/autorelease Project: yangtools  049e2e49f85525e83e13f0a4759e9126bc6edee0
BUG-4278: Improve TOUCH operation effects

The implementation of TOUCH operation would bump the subtree version of
the node even if there were no children coming in, resulting in needless
promotion of LazyContainerNodes to MaterializedContainerNodes.

If a TOUCH operation does not result in any of the children being
modified, the subtree version should not be bumped, since the subtree
was not changed. This should be the case if the user-specified children
are empty and also if the all nodes end up not modifying the data tree.

Detect both conditions and reuse the original metadata node instead of
replacing it -- this preventing LazyContainerNode promotion.

Change-Id: Icdbde02e7c3f503a9ada28b58fd27e3e3a3ef812
Signed-off-by: Robert Varga <rovarga@cisco.com>