Updated releng/autorelease Project: yangtools  f57567345ea5fa7d38179889e2748003a38c384f
BUG-4288: check for existing child modification

When we are issuing a merge on a previously-written node we need to make
sure the written node's data is pushed down to child nodes. This could
have ended up overwriting nodes which have occured under and after the
write being replaced. Guard against this mishap with an explicit check.

Change-Id: I2425dc758b8099de349641cd3bef9949cfccf57c
Signed-off-by: Robert Varga <rovarga@cisco.com>