Updated releng/autorelease Project: vtn  9a04c7f51dee84fb68148494aace7b3322c56b1b
Wait for completion of network topology detection on port link-up.

Network topology detection by LLDP will be initiated when link-up is
detected on a switch port. In that case the VTN Manager should not process
PACKET_IN and PACKET_OUT on that port until the topology detection
completes.

  * vtn-config: node-edge-wait was renamed to topology-wait.

Other changes:

  * Refined JUnit test for VtnNodeEvent and VtnPortEvent.

Change-Id: Ie3319e2ab84bb2bc7267b9288d6cc55f468f9802
Signed-off-by: Shigeru Yasuda <s-yasuda@da.jp.nec.com>