Updated releng/autorelease Project: aaa  05bd79dd050232f302dfc1d535d8fba81042a8e4
Introduction of the odl-aaa-shiro feature

Introduces the odl-aaa-shiro feature, based on the aaa-shiro bundle.  This
bundle is built separately from the rest of the AAA subsystem, and can stand
independent of other code in AAA.  Without appropriate changes to restconf
in the netconf project, this change is not usable in the ODL project.

This sets the AAA scaffolding for aaa-shiro integration.  This patch can
be merged without affecting other code in the aaa or restconf projects.  It
is the start to migrating from the custom AAA solution to one that is built
on Apache Shiro.  This patch supports 1.X versions of JAX-RS at this time,
and will later provide functionality to support the 2.X API.

This introduction of code adds/allows the following Realms:
1) ODLJndiLdapRealm
2) INI file based
3) JDBCRealm

There is stub support for:
1) RadiusRealm
2) TACACSRealm
But these have not yet been implemented.  They stand as a Proof of Concept
as to how easy it is to create different AAA backends to authenticate against.

Corresponding test methods are present for all testable methods in classes.
This is the first of many incremental patches to help migrate AAA to Shiro.
This particular patch will not actually affect the current implementation.

Change-Id: Ifbdec290d87434279882295c24e35412b641154f
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>